+++ package.json
@@ -14,21 +14,21 @@
   },
   "dependencies": {
     "@ai-sdk/react": "2.0.0-beta.6",
-    "@radix-ui/react-alert-dialog": "^1.1.14",
-    "@radix-ui/react-dialog": "^1.1.14",
-    "@radix-ui/react-dropdown-menu": "^2.1.15",
+    "@radix-ui/react-alert-dialog": "^1.1.15",
+    "@radix-ui/react-dialog": "^1.1.15",
+    "@radix-ui/react-dropdown-menu": "^2.1.16",
     "@radix-ui/react-separator": "^1.1.7",
     "@radix-ui/react-slot": "^1.2.3",
-    "@radix-ui/react-tooltip": "^1.2.7",
-    "@tailwindcss/vite": "^4.1.11",
+    "@radix-ui/react-tooltip": "^1.2.8",
+    "@tailwindcss/vite": "^4.1.13",
     "@types/react-syntax-highlighter": "^15.5.13",
     "ai": "5.0.0-beta.6",
-    "ai-latest": "npm:ai@^5.0.14",
+    "ai-latest": "npm:ai@^5.0.44",
     "class-variance-authority": "^0.7.1",
     "classnames": "^2.5.1",
     "clsx": "^2.1.1",
     "date-fns": "^4.1.0",
-    "dexie": "^4.0.11",
+    "dexie": "^4.2.0",
     "dexie-react-hooks": "^1.1.7",
     "fast-deep-equal": "^3.1.3",
     "lucide-react": "^0.525.0",
@@ -37,31 +37,31 @@
     "react-dom": "^19.1.1",
     "react-error-boundary": "^6.0.0",
     "react-markdown": "^10.1.0",
-    "react-router": "^7.7.1",
-    "react-syntax-highlighter": "^15.6.3",
+    "react-router": "^7.9.1",
+    "react-syntax-highlighter": "^15.6.6",
     "remark-gfm": "^4.0.1",
     "sonner": "^2.0.7",
-    "swr": "^2.3.4",
+    "swr": "^2.3.6",
     "tailwind-merge": "^3.3.1",
-    "tailwindcss": "^4.1.11",
+    "tailwindcss": "^4.1.13",
     "usehooks-ts": "^3.1.1",
     "uuid": "^11.1.0",
     "zod": "^3.25.76"
   },
   "devDependencies": {
-    "@eslint/js": "^9.32.0",
-    "@types/node": "^24.2.0",
-    "@types/react": "^19.1.9",
-    "@types/react-dom": "^19.1.7",
+    "@eslint/js": "^9.35.0",
+    "@types/node": "^24.4.0",
+    "@types/react": "^19.1.13",
+    "@types/react-dom": "^19.1.9",
     "@vitejs/plugin-react": "^4.7.0",
-    "eslint": "^9.32.0",
+    "eslint": "^9.35.0",
     "eslint-plugin-react-hooks": "^5.2.0",
     "eslint-plugin-react-refresh": "^0.4.20",
-    "globals": "^16.3.0",
+    "globals": "^16.4.0",
     "prettier": "^3.6.2",
-    "tw-animate-css": "^1.3.6",
+    "tw-animate-css": "^1.3.8",
     "typescript": "~5.8.3",
-    "typescript-eslint": "^8.38.0",
-    "vite": "^7.0.6"
+    "typescript-eslint": "^8.43.0",
+    "vite": "^7.1.5"
   }
 }
